Fixed VC++ warnings
svn-id: r7652

+++ scumm/gfx.cpp
@@ -3106,7 +3106,7 @@ void Scumm::createSpecialPalette(int16 from, int16 to, int16 redScale, int16 gre
 
 		searchPtr = palPtr + from * 3;
 		bestResult = 32000;
-		currentIndex = from;
+		currentIndex = (byte) from;
 
 		for (j = from; j < to; j++) {
 			int ar = (*searchPtr++);
